+++ contracts/tranches/interfaces/IAccounting.sol
@@ -14,8 +14,8 @@ interface IAccounting is ICDOComponent, IAprPairFeedListener {
         uint256 srtAssetsOut
     ) external;
 
-    function totalAssetsT0() external view returns (uint jrtNavT0, uint srtNavT0, uint reserveNavT0);
-    function totalAssets (uint256 navT1) external view returns (uint jrtNavT1, uint srtNavT1, uint reserveNavT1);
+    function totalAssetsT0() external view returns (uint256 jrtNavT0, uint256 srtNavT0, uint256 reserveNavT0);
+    function totalAssets (uint256 navT1) external view returns (uint256 jrtNavT1, uint256 srtNavT1, uint256 reserveNavT1);
     function totalReserve () external view returns (uint256);
     function reduceReserve (uint256 amount) external;
